# Adds NEON versions of cos sin and tan.