Adds NEON versions of cos sin and tan. The reason they are in the same commit is that they are calculated by the same underlying function.

RUN_VOLK_TESTS: volk_32f_cos_32f(131071,1)
generic_fast completed in 10.0344 ms
generic completed in 3.34385 ms
neon completed in 1.44828 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 3.34994 ms
neon completed in 1.46552 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_tan_32f(131071,1)
generic completed in 9.04345 ms
neon completed in 1.99319 ms
Best aligned arch: neon
Best unaligned arch: neon