@hcab14 2019-12-11
As far as I understand, prefetching beyond the size of the array does not do any harm. Since in some cases prefetch helps I would keep it.

Also found this reference
https://developer.arm.com/docs/100616/latest/functional-description/level-1-memory-system/data-prefetching

where 3 different ARM prefetch instructions are mentioned: PLD, PLDW, PLI, PRFM. Might be worth to understand their difference and if they can help.