@michaelld 2019-12-11
Right now `VOLK_PREFETCH` for GCC and Clang uses the compiler directive `__builtin_prefetch`, which according to < https://gcc.gnu.org/onlinedocs/gcc/Other-Builtins.html >:
```
Data prefetch does not generate faults if addr is invalid, but the address expression itself must be valid. For example, a prefetch of p->next does not fault if p->next is not a valid address, but evaluation faults if p is not a valid address.
```
So since the pointer value is valid, it doesn't matter (at least for GCC) whether the dereferenced pointer is valid. I'd guess the same holds for Clang. Hence, let's keep the prefetching.