@balister 2019-12-11
We should confirm the prefetches have benefit and if not remove them. Less code, less to go wrong.