@jdemel 2019-12-12
@ast could you do a quick benchmark?
I hope this does help though. Synthetic benchmarks are always difficult to apply to real world problems.
Actually, I'd be in favor of leaving `VOLK_PREFETCH` in. Even if we can't confirm an immediate performance gain. Mostly because we can only run synthetic benchmarks and we could already confirm, that a prefetch does not cause any harm. Though, we could add a warning comment to that line in code. If it ever fails, we have a reminder why.