@ast 2019-12-12
Here's some output of `volk_test_all volk_32f_sin_32f` there *might* be a tiny effect of `VOLK_PREFETCH` but to conclude anything I think we would need a much longer running benchmark presenting mean and sd.

It would be nice if it could output data for GNUPlot also...

## With prefetch

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 8.61041 ms
neon completed in 3.69726 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 8.61528 ms
neon completed in 3.70294 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 3.42371 ms
neon completed in 1.49339 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 8.49853 ms
neon completed in 3.65365 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 3.40036 ms
neon completed in 1.46633 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 8.5379 ms
neon completed in 3.65135 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 3.40914 ms
neon completed in 1.46822 ms
Best aligned arch: neon
Best unaligned arch: neon



## Without

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 8.64338 ms
neon completed in 3.76444 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 8.63676 ms
neon completed in 3.65867 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 8.48042 ms
neon completed in 3.65633 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 3.43347 ms
neon completed in 1.52261 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 8.52323 ms
neon completed in 3.65681 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 8.58769 ms
neon completed in 3.65761 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 3.40534 ms
neon completed in 1.49404 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1)
generic completed in 8.58091 ms
neon completed in 3.70993 ms
Best aligned arch: neon
Best unaligned arch: neon