@hcab14 2019-12-12
Instead of ctest you can also run `volk_profile -R volk_32f_sin_32f` which uses more than one iteration.

Your timings show that the CPU frequency is changing. If you are using an RPi3 this can be avoided by adding the following two lines to `/boot/config.txt`
```
core_freq=400
core_freq_min=400
```