@ast 2019-12-12
Ah ok, of course.

## With prefetch 

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1987)
generic completed in 6704.08 ms
neon completed in 2902.71 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1987)
generic completed in 6709.87 ms
neon completed in 2922.11 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1987)
generic completed in 6710.97 ms
neon completed in 2905.6 ms
Best aligned arch: neon
Best unaligned arch: neon

## Without

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1987)
generic completed in 6655.48 ms
neon completed in 2906.17 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1987)
generic completed in 6723.19 ms
neon completed in 2903.95 ms
Best aligned arch: neon
Best unaligned arch: neon

RUN_VOLK_TESTS: volk_32f_sin_32f(131071,1987)
generic completed in 6706.33 ms
neon completed in 2903.89 ms
Best aligned arch: neon
Best unaligned arch: neon