@michaelld 2019-12-12
Looks like there is nominal performance gain from the prefetch. I'm still in favor of leaving the `VOLK_PREFETCH` in.